Split up PoC compiler into independent modules
