+++ xircuits/compiler/__init__.py
@@ -1,0 +1,1 @@
+from .compiler import compile
